fix: change parameter datatype

+++ app/src/main/java/com/example/mycommish/feature/prize/domain/repository/PrizeRepo.kt
@@ -5,7 +5,7 @@ import kotlinx.coroutines.flow.Flow
 
 interface PrizeRepo {
     fun getAllPrizes(): Flow<List<PrizeEntity>>
-    fun getSinglePrizeById(id: Int): Flow<PrizeEntity?>
+    fun getSinglePrizeById(id: Long): Flow<PrizeEntity?>
     suspend fun insertPrize(prizeEntity: PrizeEntity)
     suspend fun updatePrize(prizeEntity: PrizeEntity)
     suspend fun deletePrize(prizeEntity: PrizeEntity)
